+++ fake/lms/celery.py
@@ -1,0 +1,4 @@
+from celery import Celery
+
+
+APP = Celery('proj')